refactor /app/welcome into components
